Time: 82 ms (34.45%) | Memory: 158.1 MB (49.28%) - LeetSync

+++ 3974-xor-after-range-multiplication-queries-i/xor-after-range-multiplication-queries-i.cpp
@@ -4,7 +4,7 @@ public:
         for(int i = 0;i<q.size();i++){
             int idx = q[i][0];
             while(idx<=q[i][1]){
-                nums[idx] = ((long long)nums[idx] * q[i][3]) % (1000000007);
+                nums[idx] = ((long)nums[idx] * q[i][3]) % (1000000007);
                 idx+=q[i][2];
             }
         }
